+++ projects/iad-interface-admin/form/src/lib/dynamic-form/dynamic-form.module.ts
@@ -14,17 +14,16 @@ import { IadFormDateComponent } from './components/iad-form-date.component';
 import { IadFormFileComponent } from './components/iad-form-file.component';
 import { IadFormInputComponent } from './components/iad-form-input.component';
 import { IadFormNumberComponent } from './components/iad-form-number.component';
-import { IadDropdownGroupComponent } from './components/iad-dropdown-group.component';
+import { IadFormGroupComponent } from './components/iad-form-group.component';
 import { IadFormDateTimeComponent } from './components/iad-form-date-time.component';
 import { IadFormTextareaComponent } from './components/iad-form-textarea.component';
 import { FileUploadComponent } from './file-upload/file-upload.component';
 import { IadFormCheckboxComponent } from './components/iad-form-checkbox.component';
-import { IadFormMultiSelectComponent } from './components/iad-form-multi-select.component';
-import { IadFormSelectionDropdownComponent } from './components/iad-form-selection-dropdown.component';
+import { IadFormDropdownComponent } from './components/iad-form-dropdown.component';
 import { IadFormRichEditorComponent } from './components/iad-form-rich-editor.component';
 import { IadFormChipsComponent } from './components/iad-form-chips.component';
 import { IadFormTranslateInputComponent } from './components/iad-form-translate-input.component';
-import {IadFormAutoCompleteInputComponent} from './components/iad-form-auto-complete-input.component';
+import { IadFormAutoCompleteInputComponent } from './components/iad-form-auto-complete-input.component';
 
 @NgModule({
     imports: [CalendarModule, InputMaskModule, PanelModule, IadSharedModule, IadPrimengModule, FontAwesomeModule],
@@ -36,13 +35,12 @@ import {IadFormAutoCompleteInputComponent} from './components/iad-form-auto-comp
         IadFormDateComponent,
         IadFormInputComponent,
         IadFormNumberComponent,
-        IadDropdownGroupComponent,
+        IadFormGroupComponent,
         IadFormFileComponent,
         IadFormDateTimeComponent,
         IadFormTextareaComponent,
         IadFormCheckboxComponent,
-        IadFormMultiSelectComponent,
-        IadFormSelectionDropdownComponent,
+        IadFormDropdownComponent,
         IadFormRichEditorComponent,
         IadFormChipsComponent,
         IadFormTranslateInputComponent,
@@ -52,13 +50,12 @@ import {IadFormAutoCompleteInputComponent} from './components/iad-form-auto-comp
         IadFormDateComponent,
         IadFormInputComponent,
         IadFormNumberComponent,
-        IadDropdownGroupComponent,
+        IadFormGroupComponent,
         IadFormFileComponent,
         IadFormDateTimeComponent,
         IadFormTextareaComponent,
         IadFormCheckboxComponent,
-        IadFormMultiSelectComponent,
-        IadFormSelectionDropdownComponent,
+        IadFormDropdownComponent,
         IadFormRichEditorComponent,
         IadFormChipsComponent,
         IadFormTranslateInputComponent,